requirements.txt
aiohttp==3.8.4
aiohttp-sse-client==0.2.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Django==3.2.19
django-cors-headers==3.14.0
django-environ==0.10.0
django-model-utils==4.3.1
django-polaris==2.3.7
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
mnemonic==0.20
multidict==6.0.4
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stellar-base-sseclient==0.0.21
stellar-sdk==8.2.1
toml==0.10.2
typeguard==2.13.3
urllib3==2.0.3
whitenoise==5.3.0
yarl==1.9.2
